Amélioration

+++ ghannoum_toihir/analyseur_syntaxique.h
@@ -9,7 +9,7 @@
 
 void build_tree_analysis(grammar G, table T, char* chaine);
 int search_state_table(table t, int etatDepart, char caractereArechercher);
-void print_tree_analysis(char* chaine, int tailleChaine, char* pile, int taillePile, int i);
+void print_tree_analysis(char* chaine, int tailleChaine, char* pile, int taillePile, size_t i);
 void print_derivations(int* reductions, int tailleReductions, grammar G);
 
 #endif
